Initial Cleanup
It's my current belief that all of the .txt files in this repo were either output files from experiments or circuits that were used for testing and benchmarking ooorrrr... Honestly I'm not certain, the bulk of the clutter came from .txt files so it's tough to parse out all of it. Additionaly, all of the pictures used in writing the FQsun paper, as well as some presentations, have been removed because they're not applicable to emulation.

The other main element removed was the benchmarking code, as well as a lot of jupyter notebooks and python scripts used for things like testing. All of this is bloat as far as I'm concerned.

Next objectives:
- Continue cleanup
- Run small and simple qiskit simulations on simple quantum circuits.