Fixing qoop submodule
Discovered that the qoop folder, which is included as python code in several other files, was intended to be a submodule. However the original submodule doesn't seem to exist anymore, so I've found what I believe is the correct repo from the same author as the main body of code and added it back in as a submodule.