Experimenting with Python
Added notes into README as a roadmap for myself.

Decided to experiment with the Qsun python package by desinging some simply quantum circuits. Starting off with a basic Deutsch-Jozsa running on a parity oracle. Early findings: the Qsun package is put together pretty cleanly but it's not the post pythonic thing I've ever encountered. For example, the "circuit" variable paradigm is very atomic: it's seeming difficult to build Deutsch-Jozsa in such a way as to be agnostic of the oracle we're interested in using. So the entire circuit has to be designed start to finish. I will see if I'm wrong, but if I'm not I will attempt to pythonify the package a bit.